kerasioimg/requirements.txt
tensorflow==2.5
imutils==0.5.4
opencv-python==4.5.1.48
matplotlib==3.3.4
pandas==1.1.5
imageio==2.9.0
pytube
flask
gunicorn
WebTest
tensorflow_hub
scikit-image
lime
matplotlib
google-cloud-storage==1.38.0